requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
schedule==1.2.1
requests==2.31.0
sendgrid==6.11.0 